Re-write of RtpPacketQueue.cpp, which was not correctly handling:
1. Out of order data
2. Jumps in data
3. RTP sequence number rollover
4. Maintaining queue depth (it was simply nuking the entire queue any time it grew too large instead of trimming it)

Use of the std::map structure seems like a good choice at first, but it is problematic for the first three issues. a std::list is a better choice, given that A) most inserts are going to be at or near the front of the list, minimizing linear search cost and B) traversing a list of 100 things is trivial anyway.  It's more simple and doesn't have some of the above corner cases.

Also change the dataPacket size to 1500; it's possible to have video packets that are larger than 1200 bytes, particularly for keyframes.

Lastly, change ExternalOutput not to write any data while the audio and video queues have fewer than 15 items in them.  This is to allow a window of time for packet re-ordering to occur.  I want to make this more sophisticated in the near future.