Fix so we don't record corrupted frames.
When there's packet loss, there's a possibility that we end up constructing a corrupted frame and hand that to libav.  The resulting corruption is much worse than if we had just skipped the frame entirely.  This change also fixes audio handling (it was incorrectly computing header size) and optimizes it slightly so we're not doing an unnecessary memcpy.

One side affect of this is we have to examine the VP8 header before we copy it into our unpackage buffer.  This also ends up optimizing stuff slightly though.

Also, zap input processor member variable and audio buffer; we don't need either of these anymore.