Import the WebRTC VCM based video mixer code from Qi.
Additional modifications on the build script and directory structure is made.
TBD: the BufferManager should be placed into woogeen_base, but we now place
it in mcu temporarily because it relies on boost version above 1.53, which is
not always available for our current building system. We'll move it to the
correct place once we have a solution on the dependency.