Add the stream when ice negotiation finishes.
When it takes longer to send the 'ok' message over socket.io than it
takes to move the ice connection into the complete state, we end up
failing to notify everyone else in the room that there is another
stream.

Instead of waiting for the 'ok' message, create the stream when the ice
connection moves into the complete state. This eliminates the race
condition and ensures that a connected stream is always fully published.

The other side effect here is that the 'ok' message is no longer needed.
I left it for compatibility, but it can easily be removed.