Get rid of OneToManyProcessor from the Mixer.
The Mixer should be at the same level as OneToManyProcessor, and manage
the subscribers as MediaSinks.
This way the MCU module won't rely on the erizo module, except for implementing
some public erizo interfaces (i.e., interfaces defined in MediaDefinitions.h).
In the future the publisher/subscriber management should be further abstracted.