Set remote/source SSRC for the VCMInputProcessor.
When building the RTCP packet, the rtp_rtcp module sometimes doesn't need
explicit remote/source SSRC setting because it can get the information
from the received RTP packet (for example when building the RR), but in
some other cases (like building NACK) it requires the remote/source SSRC
being explicitly setted.
With this, the MCU is able to send correctly formatted NACK to the publishers.