Introduce the initial implementation of Out-of-process Mixer.
Use the RawTransport to do inter-process data transfer between the publishers and the mixer.

Although currently it's a POC, it works! To try it out, set
"config.erizoController.outOfProcessMixer = true" and "config.erizoAgent.maxProcesses = 17".
(the maxProcesses should be  >=2.)

Please note that currently as a POC,

  1) The OOP Mixer feedback is not delivered to the publisher, so that you may observe
     QoS issue;
  2) There's no notification to the OOP Mixer if a publisher leaves the room, so that
     the OOP Mixer is not able to adjust the video layout accordingly;
  3) A/V sync is not enabled for the OOP Mixer.

Those issues should be fixed if we add

  1) control messages (we need to define them) between the publishers and the OOP Mixer;
  2) data send channel from the OOP Mixer to the publishers.

Also, MediaSourceConsumer and Gateway interfaces could be further improved to make them
more clear for their roles.