Fix a bug in adding sources to the mixer.
In order to workaround the current OOP Mixer limitation we have added
a mechanism to add the mixer source on-demand. There was a random bug
that different threads will attempt to add the same source to the mixer.
It's fixed for now, but in the future we won't allow adding source on-demand
at all once the OOP Mixer is able to explicitly invoke addSource.