Refine the VideoFrame interface.
Introduce VideoFrameEncoder and split the original VideoFrameProcessor
into VideoFrameCompositor and VideoFrameEncoder.

For hardware, HardwareVideoMixer is now both a VideoFrameCompositor and
a VideoFrameEncoder. We may think about spliting it in the future.

For software, a faked VideoFrameEncoder is implemented to dispatch the
raw frames generated from the VideoFrameCompositor (SoftVideoCompositor),
and the real encoding work is still performed in VCMOutputProcessor. We
may extract the encoder part from the VCMOutputProcessor later.

TODO is marked to provide VideoFrameDecoder interface as well.