Remove EncodedVideoFrameCompositor and introduce VideoFrameMixer concept.
So a VideoFrameMixer accepts frames from multiple inputs and outputs frames
in different formats.
For hardware implementation of VideoFrameMixer, it's now naturally mapped to
the VCSA hardware mix engine.
For software implementation of VideoFrameMixer, it's now tries to de-couple
the decoder, compositor and encoder stages by employing VideoFrameDecoder,
VideoFrameCompositor and VideoFrameEncoder.