Don't perform further operations on a WebRtcConnection if we receive the CONN_FINISHED event.
It's fragile - the synchronization between the node.js control thread and the native work threads.
Currently we use async_event mechanism to notify the node.js control thread about the connection status
from the native layer. However, the order in which the events arrive may be different than that when the
events are sent out. There could be a case that during the connection establishment process the user
disconnects. In such case the node.js layer may receive both a CONN_FINISHED and another normal connection status.