Add an option to prompt an asynchronous event from native to JS control layer.
Some events from native (for example, the WebRtcConnection destroy event) should be
notified to the JS control layer as soon as possible so that he JS controller won't
ask the WebRtcConnection to do something if it receives some other events from other
working threads later. One example is that the WebRtcConnection is destroyed before it's
ready for transporting media (in that case the controller may ask the WebRtcConnection
to getLocalSdp if it doesn't know that the WebRtcConnection is destroyed).