Update RTP packet queue to be based on time, not packet count
The problem with basing it on packet count is it makes sync between our audio and video tracks difficult because # of packets has nothing to do with the duration of the content.  That makes it hard to have a deep queue, but also not cause libav to freak out when there's significant deviation between incoming audio and video timestamps.

Change it so our queue depth is based on the duration of samples in the queue rather than an arbitrary packet count.  Also update our tests to reflect this change