Support different audio codec outputs for different participants.
The preferred sent audio codec for a specific participant will be set
based on the preferrence of SDP negotiation.
This means that the audio mixer is able to output different audio codec
formats, similar to the video mixer. For example if we enable ISAC (currently
disabled in our SDP) in server then Chrome will send and receive ISAC
encoded audio, and Firefox (which doesn't support ISAC) will send and receive
PCMU encoded audio.