Fix a synchronization issue in WebRtcConnection destruction.
Firstly, copy the removed subscribers to a local vector before removing
them from the subscribers map, then we release the lock for the map.
This allows us to destroy the subscribers after releasing the map lock
which avoids potential dead-locks between the main thread (destroying
the subscribers) and the WebRTC threads (receiving and sending streams).