Support multiple mixed video stream outputs for one codec in the software mixer.
This is mainly to support the clients with different QoS capabilities,
for example Chrome supports RED/FEC and NACK while Firefox supports NACK only.
In that case we should send different mixed streams to Chrome and Firefox,
and FEC/NACK for each stream should be enabled based on the client capability
described in their offer SDP.

Currently this feature is only added/enabled in the software mixer.