Enable building and packaging together the node binary with erizoJS.
This is not only for code protection, but also for functionality.
Both erizoJS and standard node binary will statically link the openssl library,
but if we use the node binary downloaded from the public website, it may
have included a different version of openssl which may conflict with ours,
and we also cannot make sure we're using the openssl version which meets the
security requirement.
We need to make sure the node binary and erizoJS library are using the same openssl
library, and by building the node binary manually with the same openssl library path
specified is the most reasonable approach for now.