Add the implementation of A VCM based video frame encoder.
The encoder accepts the raw I420 frames and encodes them to the given format.

This can be used for software based transcoding.
The software mixer can also use this together with the EncodedVideoFrameSender, but
for now it still uses the original compound VCMOutputProcessor because of better
QoS control. In case we need to separate frame encoding/decoding from transportation,
we can use this.
This also provides more choices for the media recording implementation.

Change-Id: I0328d32cd061308b70df33cef17954d5b02c308c