Refactor the native layer Mixer and Gateway interfaces.
We're going to expose more interfaces which are specific to the Mixer,
such as the participant and video layout region management etc..
This makes it hard to still keep the Mixer look like a Gateway, though
the original design was to make the JS and binding layers more clean, and
to make it easy to connect the muxer and mixer together.

I tried to make all of the layers: native, binding and JS still as
clean/simple as possible, though some compromise has to be made.

Change-Id: Id827b10b00aa28e9a3371b7f93685b5f257db4a8