Add getRegion/setRegion API.
This set of mixer specific APIs are used to get and set
the region ID of a publisher/partcipant in the video mixer.

Do we have some guidelines on client SDK API names, parameters
and return values? Also do we have any good suggestion on improving
the JavaScript code redundancy at both the server side and the client side?

Change-Id: Iccc9ac8e49bd468e1b7668d3fad16717184c69ac