Add the API to change the bitrate of a publisher of a mixer.
Currently we only support customized bitrate change when the mixer
is available. Later we'll add the support to pure forwarding mode conference.

We now utilize the RTCP REMB feedback to change the bitrate of a publisher.
Going forward we may need to adjust the implementation if the clients (browsers)
support more MediaStream APIs like applyConstraints etc.

Change-Id: I86f34374ec0ec19bf884ad3e24beecf2e15d9d77