Use payload type instead of codec name to identify a media source codec type.
We now have too many diverged approaches to marking the codec type of a media source,
including the codec name (string), the payload type, FrameFormat and webrtc::VideoCodecType etc,
and it's a big mess when we need to do the mapping between each other.

We should try to unify the codec format description in our code.