Modify the media muxer interface.
Don't use a MediaFrameQueue to pass the frame data from the frame provider (e.g.,
the mixer, the encoder or the frame constructor, etc.) to the media muxer (e.g.,
the recorder or RTSP output), because in some cases a FrameQueue doesn't fit, for
example the RTSP audio output. Instead to make the concept simple, we can treat
the recorder and RTSP output as a FrameConsumer - the frame provider just needs
to feed it with a frame, and the consumer will do whatever it wants upon receiving
a frame (e.g., in some cases pushing it into a queue).

Also some names are changed to avoid confusions (e.g., MediaMuxing vs MediaMuxer).

Change-Id: Ie555d25bb2b6a5c26e4968b7af98a4a987ad67c8