Fix a timeout issue when adding an external input (e.g., RTSP input).
When the signaling server (i.e., erizoController) sends a command to
the media server (erizoJSController) to add an external (RTSP) input,
erizoJSController will respond to the erizoController when the underlying
media connection between the MCU and the RTSP source is established.
In some cases this may take a long time (e.g., longer than 5 seconds),
and when this happens, erizoController won't receive the response from
erizoJSController about the RTSP input being added successfully, because
the RPC timeout is set to be 5 seconds in our system.

This patch will make erizoJSController send a response to the erizoController
w/o waiting for the RTSP media connection being established to keep the RPC
connection, and when the media connection is established, it will send another
"onReady" response to erizoController so that erizoController can send onStreamAdd
messages to the room.

This is similar to the regular WebRTC publish/subscribe process, to make
an additional onReady callback for the actions which may take longer time than
the RPC timeout.