Fix a corner case if the client disconnects from the MCU before it's stream is published successfully.
We need to close the underlying stream channels in such case.
This is caused by the fact that the media stream channel establishment is an async operation.
When the signaling server handles the disconnection from a client, it may not be aware of the
underlying media channel status before it's notified (by the native media layer).