Add an interrupt mechanism to ExternalInput.
We use libav to handle the external inputs (e.g., files, RTSP inputs etc.).
Sometimes the the libav blocking APIs will get stuck (like av_read_frame),
which prevents the MCU from getting any more data from the external input,
and also brings some threading issues when the conference is being closed.

Current solution is to interrupt av_read_frame if the operation timeouts at
a given value (say, 1 second), and silently re-open the input upon the failure.
We currently do it silently without notifying the receivers and the MCU controllers,
and assume the stream (id, codec etc.) won't be changed if we re-open the input.
Later we need to revisit this place to
1) understand the root cause why av_read_frame gets stuck;
2) figure out if there's better solution to this problem.