Fix an H264 forward issue in MCU.
If the MCU will parse the H264 packet (either for mixing or for recording or for transcoding),
the original packet may be modified (by the upstream h264 receiver implementation in upstream webrtc stack).
This will impact further forwarding and processing of the same packet in MCU scenario.
Detail description is available in the code comment.