We should correctly set the remote NiceCandidate's foundation.
The foundation is the ID of a remote nice candidate, and libnice depends
on it to identify a remote candidate. If we don't set the foundation which
is carried by the IceCandidate message or SDP, all of the remote candidates
will have the same foundation (i.e., same ID) and libnice may incorrectly
mark a selected candidate pair for the WebRTC peer connection, and thus
the peer connection establishment fails.