Manage the WebRtcConnection objects lifetime explicitly at the JS controller/binding layer.
The WebRtcConnection objects are created by the JS controller/binding layer but previously
it's deleted implicitly in the native core layer, which seems very confusing and hard to
maintain. We now have the JS controller/binding layer which creates those objects to
delete them as well.

A basic reference count is added to the WebRtcConnection object at the binding layer.

Change-Id: I7b749e5ed5ec9c61adeec7b47d08918630609280
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/73346
Reviewed-by: Xian, Yuqiang <yuqiang.xian@intel.com>
Tested-by: Xian, Yuqiang <yuqiang.xian@intel.com>