Improve the time and space efficiency for assigning a free index for an input.
Time efficiency is changed from original O(N^2) to O(N);
Space efficiency is improved by using the bool/bit vector which 1) represents
each element with a single bit; 2) avoids dynamic realloaction by reserving
the capacity to be the max number of inputs from the room configuraton.

The benefit may be more obvious if the user configures the max inputs to be a huge number.

Change-Id: I79b3486fee9116484e67576e02f3fca23a719d2e
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/73678
Reviewed-by: Duan, Xiande <xiande.duan@intel.com>
Tested-by: Xian, Yuqiang <yuqiang.xian@intel.com>