Make the RawTransport buffer size configurable.
The send buffer size in the send queue will be equal to the data length to be sent;
the receiver buffer size will be specified by the user of the RawTransport when it's created.

Next step we need to create the transports using different buffer size parameters
depending on the specific usage of the transports, e.g., whether it's used to transfer
large data such as the high resolution video streams.

Change-Id: I3a8035a7ed271c59f686bdaf0d285dca92b3ca9f
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/73998
Reviewed-by: Duan, Xiande <xiande.duan@intel.com>
Tested-by: Xian, Yuqiang <yuqiang.xian@intel.com>