Fix bug #228.
The first parameter of Object.create() needs to be the prototype of the newly created object,
which should be a null or non-null object reference. It can't be undefined which will cause
the JS engine "Throws a TypeError exception if the proto parameter isn't null or an object."
(https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/create).

Change-Id: I322de1d3cfa2040f9ec20c89420cb43ed6bf22eb
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/74421
Reviewed-by: Zhu, Jianjun <jianjun.zhu@intel.com>
Tested-by: webrtctest <webrtctest@intel.com>
Reviewed-by: Xian, Yuqiang <yuqiang.xian@intel.com>