Add a new video mixing option "crop" in room configuration.
The new option is for the user to specify whether to crop the video input
to make it filled into the region for it in the mixed video.

Currently this is only valid for software mixer/compositor.
We need to check whether there's API for the hardware mixer for us to
control the behavior similarly.

Change-Id: I728a40039b8321786f5b4bf771fca953fbd830e8
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/78462
Tested-by: webrtctest <webrtctest@intel.com>
Reviewed-by: Xian, Yuqiang <yuqiang.xian@intel.com>