Fix a memory issue in RawTransport.
Using a union to hold complex type members is error-prone, because
the members are sharing the same memory, and the destructor/constructor
of the members may not work correctly in a union. From valgrind
we could see that the destructor of the members in the union is never
invoked, thus memory leak happens.

Change-Id: I838dd5e9ad3a1ff6ece6a28b257e381ecde06a67
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/90793
Tested-by: webrtctest <webrtctest@intel.com>
Reviewed-by: Duan, Xiande <xiande.duan@intel.com>
Reviewed-by: Xian, Yuqiang <yuqiang.xian@intel.com>