Fix a memory leak on CentOS.
Use boost shared_array instead of shared_ptr to hold the pointer to an array.
Although it's mentioned in Boost document that since Boost version 1.53 shared_ptr
should also be able to handle the array deletion by overloading operator[], the
behavior on CentOS 7.x (with system provided libboost 1.53.0) seems not complying
with that, and results in memory leaks.

Change-Id: I8bc87b87263fbbc02521dc5e3de0bc5a0b721573
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/94081
Reviewed-by: Duan, Xiande <xiande.duan@intel.com>
Tested-by: webrtctest <webrtctest@intel.com>
Reviewed-by: Xian, Yuqiang <yuqiang.xian@intel.com>