Fix a "stupid" bug which causes random failures in stress test.
Constructing a "scoped_lock" object without given a name means nothing at all -
nothing will be locked. And in this case it will potentially break the internal data
structure of a STL deque, for example to make the deque iterator in a wrong state
where the "current" pointer points to a place beyond the range between "first" and "last".

This is a deeply hidden bug and bothered us for a long time.

Change-Id: Ia75ed6e87a92feb1358682d53188c3351af2feed
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/95576
Tested-by: webrtctest <webrtctest@intel.com>
Reviewed-by: Xian, Yuqiang <yuqiang.xian@intel.com>