Make sure the feedback sink of the video frame constructor always valid.
We need to unset the feedback sink of the frame constructor when the feedback sink is being closed.
For example, the WebRTC connection is now both the media source and the feedback sink of the frame constructor,
and we keep the constructor alive before the WebRTC connection is destroyed because the frame constructor is
the media sink that the WebRTC connection references; but the WebRTC connection is also the feedback sink that
the frame constructor references. This cyclic reference should be broken when a node in the cycle is being removed,
and it works by adding the "breaking" operations in the destructor of each node.

Reference: https://git-amr-1.devtools.intel.com/gerrit/92422

But For the video frame constructor, the reference breaking mechanism doesn't work because it lacks something
for the media source (i.e., a WebRTC connection) to track and notify when the media source is being destroyed. So we
set the feedback source of the video frame constructor to make the underlying reference breaking mechanism work.

Change-Id: I2560e8cbed0820b9bef126cb0a65448927ec2eb7
Reviewed-on: https://git-amr-1.devtools.intel.com/gerrit/97972
Reviewed-by: Duan, Xiande <xiande.duan@intel.com>
Reviewed-by: Xian, Yuqiang <yuqiang.xian@intel.com>
Tested-by: Xian, Yuqiang <yuqiang.xian@intel.com>