Separate rtcController and quicController.
Conference sessions over QUIC used to be handled by accessController.
But a recent change added rtcController, moved some logicals out of
accessController. So we need to either have a quicController or extend
rtcController or accessController. This change adds a quicController.
However, it would be better to redesign the interface of *Controller,
and create a single controller for each conference.