Add an agent for WebTransport Support. (#113)
* Add RTCIceTransport and RTCIceCandidate.

Change-Id: If39775bf3494bdceeef44846abfb775dbd4e8a65

* Add RTCQuicTransport.

* Receive data from client.

OnData event is not complied with spec. Fin is not respected.

* Write data to QuicStream.

* Add a task runner to call QUIC APIs.

* Fix a memory leak issue.

* Move QUIC to a separate directory.

* Separate WebRTC and QUIC code.

This experimental change makes WebRTC agent running QUIC code. We'll
move QUIC to a standalone agent in future changes.

* Add signaling support for publishing a QUIC stream.

* Forwarding data through QUIC data channel.

This change adds data filed to publication and subscription.

* Adding interfaces for data pipeline in C++ level.

* QuicStream implements FrameSource and FrameDestination.

* Forward data in C++ level.

* Add documentation for how to run server with QUIC enabled.

* Update doc for CHROMIUM_HOME.

* Fix argument issue for quicConnection.js.

* Fix dependency issue during building.

Certificate is not used right now, so we don't need to initialize
SSL at this time. It removes its dependent to rtc_base.

* Copy data during forwarding.

QuicStringPiece may read freed buffer when write to transport.

* Fix a memery leak issue.

* Move QUIC transport to a standalone agent.

* Add more files to QUIC agent.

* QUIC agent can be started and stopped.

* Enable unit tests for RTCIceTransport.

* Add an implementation of P2PQuicPacketTransportInterface.

* Refactor QUIC addon.

Moving QUIC implementation to QUIC SDK.

* Implement RTCQuicTransport::listen.

* Enable write observer and read delegate.

* Implement delegate for P2PQuicTransport.

This commit also removes some unused files.

* QuicStream is able to receive data.

* Implement write for QuicStream.

* Remove an EscapableHandleScope and add dtor for QuicStream.

* Remove branch specific section for README.md.

* Modify documentation for QUIC addon.

* Remove patches for Chromium.

QUIC addon depends on QUIC SDK now. Changes for Chromium will be
performed in QUIC SDK repo.

* Restore build configuration.

We used to add QUIC as a feature for webrtc agent, but we decided
to add a new QUIC agent for QUIC connections now. This change
restores modifications for webrtc agent.

* Exclude quic-agent from full package.

Don't pack quic-agent by default because CI for QUIC SDK is not ready.
So it cannot get the binary of QUIC SDK.

* Restore changes made outside QUIC addon.

* Adding QUIC agent.

* Add QUIC subscription support.

* Recover connections.js.

* QUIC P2P multiplexing support.

* Add QUIC transport server.

* Rename QuicStream to RTCQuicStream.

* Add QUIC transport streams.

* QUIC transport server is able to receive data.

* Quic transport server is able to send data.

* Remove code for QUIC API for Peer-to-peer Connections.

We will implement QuicTransport of WebTransport in server.

* Make QUIC agent to build with node.js v12.

* Add contentsessionid event for QuicTransportStream.

This event is fired when a QuicTransportStream could be associated with
a publication or subscription. Since server side defines session ID as a
published stream ID or a subscription ID, we use `contentsessionid` here
to distingulish it from Quic session.

* Fix data is undefined.

* Fix some issues to run QuicTransportServer in QUIC agent.

* Use UUID for publication and subscription ID.

UUID is better to be encoded for QuicTransport.

This commit adds a new dependency: uuid.
https://www.npmjs.com/package/uuid.

* Parse the first 16 bytes of QuicStream as UUID.

* Participants recieve notification when a data stream is added.

* Stream info has data property.

It indicates whether this is a data stream.

* Add subscription support for QUIC agent.

* Add support to subscribe data streams.

* Linkup publications and subscriptions for data streams.

* Enhance readme.

* Enable data stream forwarding.

* Enable authentication for QUIC transport.

QUIC transport is authenticated by transport ID which is issued by
the signaling module of portal. In the future, when QUIC transport
is allowed for signaling, other authentication method should be used. A
possible one is token, which is similar to the socket.io channel we are
using today.

* Remove P2P QUIC and code cleanup.

* Don't pack QUIC agent by default.

QUIC agent depends on QUIC SDK, which is not supported by testing
system.

* Add data source for linkup.

* Add unit test for QuicTransportServer.

* Fix build issues after rebasing.

* Don't start quic-agent by default.

* Clean up.

* Close connection if no transport ID is received after 10s.

* Update QuicFactory for QUIC SDK API change.

* Address some format and naming issues.

* Port number is configurable.

* Change QUIC agent port configuration from a range to a value.

QUIC agent always uses the minimal port number configured. So the range
actually is not respected.

* QUIC agent reads certificate path from configuration file.

* Fix an issue introduced by rebasing.

A new parameter was add for roomController.publish.

* Remove unused flags from build script.

* Fix race conditions by minimizing lock usages.

* Tokens have URL for WebTransport connections.

A new property webTransportUrl is added to tokens if a QUIC agent is
running.

* Add token for WebTransport in the response of join.

No update for tests for portal because it looks like no check in or
night testing is running the test code, and some cases are failed in my
development environment.

* Authenticate WebTransport connections with tokens.

Tokens are issued by portal after joining. QUIC agent asks portal for
authentication.

* Fix issues for publications after rebasing.

* Fix publish method after rebasing.

Server can detect new streams from QUIC agent, but this change breaks
rtcController and stream events sent from server don't have "data"
property. These issues along with other subscribe issues will be fixed
in following commits.

* Fix subscribe method after rebasing.

* Separate rtcController and quicController.

Conference sessions over QUIC used to be handled by accessController.
But a recent change added rtcController, moved some logicals out of
accessController. So we need to either have a quicController or extend
rtcController or accessController. This change adds a quicController.
However, it would be better to redesign the interface of *Controller,
and create a single controller for each conference.

* Fix WebRTC agent.

Changes for adding WebTransport support broken WebRTC agent. This change
fixes these issues.

* Add back origin info.

Don't know why origin is a property of media. Add it to publish info and
subscribe info.

* Terminate a session returns a promise.

A session might be managed by rtcController or quicController, since
accessController doesn't know a session's type, try to terminate it in
either rtcControoler or quicController.

* Add a configuration item to disable WebTransport support.

* Restore v11Client.js.

Protocol version was upgraded to v1.2. We don't need to touch v1.1
protocol handler.

* Add a guide for WebTransport.

* Clean up.

* Address comments.