Implement multiple input sync
The sync feature is only enabled for SRT streaming input at this moment

Each SoftInput maintains a frame queue with max size 30. If the frame
queue is full, it means this input is out of sync too much. We disable
sync on this input, and fallback to non sync mode input.

Sync to latest frame
Example:
    frame-queue-1 [t3 t2 t1 t0]
    frame-queue-2 [t4 t3 t2 t1]
    frame-queue-3 [t5 t4 t3 t2]
    the common sync timestamp range is [t3 t2], and we will sync to t3