Don't require os and runtime info. (#980)
Requires as less info as possible.

In P2P mode, instead of using browser version to determine whether a
feature is supported by remote side or not, we use capability map to
indicate whether a feature is supported.

This change doesn't update protocol version because it's compatible with
old protocol.